fix(trusty-mpm): require the `base-` separator, and report the roster on doctor's failure path
Review findings on the #4589/#4588 fix.

HIGH — the foundation-file rule read `starts_with("base")` with no separator
while its docs and its test name both said "`BASE-*` file name". Under that
rule `baseline-analyzer`, `base64-decoder`, `basecamp-sync` and
`based-reviewer` were all silently deleted from the roster: the same "a rule tm
controls eats an agent tm did not author, and no asset edit can fix it" failure
the frontmatter `role:` rule was removed for. The failure had moved, not gone.

The predicate is now `is_foundation_file`, a single shared definition requiring
`base-`. The asset-level gate calls it instead of keeping its own copy — a
private duplicate of the rule is the same drift this PR exists to remove.
`near_miss_base_names_are_not_treated_as_foundation_files` pins both
directions and, under the one-character mutation back to `"base"`, fails with
exactly the reported set: only `database-migrator` and `engineer` survive.

Every excluded file is now logged at debug with its directory. A silent
deletion is what made this undiagnosable; an operator whose `base-…` agent
vanished had no thread to pull.

MEDIUM — `check_agents`'s empty-deploy-tier `Fail` returned before the roster
was resolved, so it reported neither a count nor `unknown` while
`resolve_roster` was returning 42. An empty deploy tier is not an empty roster:
the project and generic `~/.claude/agents` tiers are independent of it. The
roster is resolved before the branch, so every status carries it.

The scanner's tests move to `delegation_authority_tests.rs` (#4318's pattern) —
the new near-miss test pushed the inline module past the 500-SLOC cap.

Refs #4589
Refs #4588
